[chore][pkg/stanza] Cleanup namedpipe input operator files
