+++ pkg/stanza/operator/input/namedpipe/config_all.go
