+++ pkg/stanza/operator/input/namedpipe/input_nonlinux.go
