+++ pkg/stanza/operator/input/namedpipe/input_test.go
